extract custom proposal function
